pyproject.toml
[project]
name = "qwen-image-edit-streamdiffusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.48.2",
    "diffusers>=0.36.0",
    "fastapi>=0.115.0",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "peft>=0.18.0",
    "torch",
    "torchvision",
    "triton",
    "transformers>=4.57.3",
    "uvicorn>=0.32.0",
]

[tool.uv]
prerelease = "allow"

[[tool.uv.index]]
name = "pytorch-nightly"
url = "https://download.pytorch.org/whl/nightly/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-nightly" }
torchvision = { index = "pytorch-nightly" }
triton = { index = "pytorch-nightly" }

